requirements.txt
# Core dependencies for LLMOCR project

# Data processing
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0
tqdm>=4.62.0

# Dataset management
datasets>=2.0.0
huggingface-hub>=0.10.0

# Benchmarking and metrics
python-Levenshtein>=0.12.0
jiwer>=2.3.0

# Configuration
pyyaml>=6.0

# HTTP requests (for downloads)
requests>=2.27.0

# Model training (optional - install based on your model choice)
# Uncomment the framework you're using:

# For Transformers (TrOCR, etc.)
# transformers>=4.25.0
# torch>=1.13.0
# peft>=0.2.0  # For LoRA training

# For EasyOCR
# easyocr>=1.6.0

# For PaddleOCR
# paddlepaddle>=2.4.0
# paddleocr>=2.6.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
